Display marker title and image until next marker or end (chapters)
Previously a marker would just be shown at a single point in the video.
With this change, the marker title (and image if the preview-thumbnails
plugin is used) is displayed for the whole duration between the current
markers timestamp and the timestamp of the next marker of the end of the
video.